Add `-D` debugging command-line argument and improve diagnostics

+++ framework/bug-mining/analyze-project.pl
@@ -30,7 +30,7 @@ analyze-project.pl -- Determine all suitable candidates listed in the active-bug
 
 =head1 SYNOPSIS
 
-analyze-project.pl -p project_id -w work_dir -g tracker_name -t tracker_project_id [-b bug_id]
+analyze-project.pl -p project_id -w work_dir -g tracker_name -t tracker_project_id [-b bug_id] [-D]
 
 =head1 OPTIONS
 
@@ -57,7 +57,12 @@ commons-lang project is LANG.
 =item B<-b C<bug_id>>
 
 Only analyze this bug id. The bug_id has to follow the format B<(\d+)(:(\d+))?>.
-Per default all bug ids, listed in the active-bugs csv, are considered.
+By default all bug ids, listed in the active-bugs csv, are considered.
+
+=item B<-D>
+
+Debug: Enable verbose logging and do not delete the temporary check-out directory
+(optional).
 
 =back
 
@@ -109,7 +114,7 @@ use DB;
 use Utils;
 
 my %cmd_opts;
-getopts('p:w:g:t:b:', \%cmd_opts) or pod2usage(1);
+getopts('p:w:g:t:b:D', \%cmd_opts) or pod2usage(1);
 
 pod2usage(1) unless defined $cmd_opts{p} and defined $cmd_opts{w}
                     and defined $cmd_opts{g} and defined $cmd_opts{t};
@@ -119,6 +124,7 @@ my $BID = $cmd_opts{b};
 my $WORK_DIR = abs_path($cmd_opts{w});
 my $TRACKER_ID = $cmd_opts{t};
 my $TRACKER_NAME = $cmd_opts{g};
+$DEBUG = 1 if defined $cmd_opts{D};
 
 # Check format of target version id
 if (defined $BID) {
@@ -195,7 +201,7 @@ foreach my $bid (@ids) {
     _add_row(\%data);
 }
 $dbh->disconnect();
-system("rm -rf $TMP_DIR");
+system("rm -rf $TMP_DIR") unless $DEBUG;
 
 #
 # Check size of src diff, which is created by initialize-revisions.pl script,
